Add Zsh default shell setting to setup.sh
The setup.sh script now checks if 'zsh' is present in the Brewfile.
If 'zsh' is found, it will attempt to set Zsh as the default shell
using OS-specific commands (dscl for macOS, chsh for Linux/Fedora).
A message is added to inform the user to restart the terminal for changes to take effect.